print_daily_hours: Correctly find last week's time break-down
I was depending on `ls -t` to find the most recent file in the archive,
but this fell over when I went back to correct (ie. edit) an older file.

I save the time break-down files with a sortable filename of YYMMDD.txt,
so I can just sort them to find the latest one by filename.